Separate parallelize and generate-parallel flags.
The former controls the Pluto option to generate schedules better
amenable to parallelization. The latter indicates whether to actually
emit parallel loops. This allows to exercise the scheduler without
actually running the result in parallel and circumvents the not yet
implemented parts of code generation. MLIR is able to parallelize loops
independently.